Fix getting currently connected dapp info
Add a reusable hook that will handle finding
a browser tab currently connected to the wallet
to make it easier to display info about connected dapp.
Don't use  `lastFocusedWindow` argument so window
popup can use the same hook for finding connected dapp.
