+++ ui/components/DAppConnection/DAppConnection.tsx
@@ -1,52 +1,15 @@
-import { PermissionRequest } from "@tallyho/provider-bridge-shared"
-import { selectAllowedPages } from "@tallyho/tally-background/redux-slices/selectors"
-import { browser } from "@tallyho/tally-background"
-import React, { ReactElement, useCallback, useEffect, useState } from "react"
-import { useBackgroundSelector } from "../../hooks"
+import React, { ReactElement } from "react"
 import ActiveDAppConnection from "./ActiveDAppConnection"
 import DAppConnectionDefaultToggle from "./DAppConnectionDefaultToggle"
+import { useDappPermision } from "../../hooks/dapp-hooks"
 
 export default function DAppConnection(): ReactElement {
-  const [isConnectedToDApp, setIsConnectedToDApp] = useState(false)
-  const [currentPermission, setCurrentPermission] = useState<
-    PermissionRequest | undefined
-  >(undefined)
-
-  const allowedPages = useBackgroundSelector(selectAllowedPages)
-
-  const initPermissionAndOrigin = useCallback(async () => {
-    const { url } = await browser.tabs
-      .query({
-        active: true,
-        lastFocusedWindow: true,
-      })
-      .then((tabs) =>
-        tabs[0] ? tabs[0] : { url: "", favIconUrl: "", title: "" },
-      )
-    if (!url) return
-
-    const { origin } = new URL(url)
-
-    const allowPermission = allowedPages.find(
-      (permission) => permission.origin === origin,
-    )
-
-    if (allowPermission) {
-      setCurrentPermission(allowPermission)
-      setIsConnectedToDApp(true)
-    } else {
-      setIsConnectedToDApp(false)
-    }
-  }, [allowedPages, setCurrentPermission])
-
-  useEffect(() => {
-    initPermissionAndOrigin()
-  }, [initPermissionAndOrigin])
+  const { isConnected, currentPermission, allowedPages } = useDappPermision()
 
   return (
     <section>
       <ActiveDAppConnection
-        isConnectedToDApp={isConnectedToDApp}
+        isConnectedToDApp={isConnected}
         currentPermission={currentPermission}
         allowedPages={allowedPages}
       />
